rename : api 경로 수정 (#57)
